Added Canvas Button
I have a Button Class that inherits from a UIObject Class and then an
instance of that button class working.  It detects mousovers and it
detects clicks and calls an alert.  Going to abandon this branch for now
and try to do the same with a bootstrap overlay on top of the canvas.